requirements.txt
pydoll-python>=2.21.0
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
